Refactor codebase into modular structure
- Moved Skills related components, services, and types to `src/modules/skills/`
- Moved Grad/Education related components, hooks, and types to `src/modules/grad/`
- Moved Job related components, services, and types to `src/modules/job/`
- Moved Career, Admin, and Resume specific files to their respective modules
- Updated `src/types/index.ts` to export types from new module locations
- Updated all imports to reflect the new structure
- Verified build and tests pass
