+++ src/modules/admin/services/adminService.ts
@@ -1,5 +1,5 @@
 
-import { supabase } from './supabase';
+import { supabase } from '../../../services/supabase';
 
 export interface UsageOutlier {
     user_id: string;
