+++ src/types/index.ts
@@ -1,7 +1,7 @@
-export * from './resume';
-export * from './job';
+export * from '../modules/resume/types';
+export * from '../modules/job/types';
 export * from './analysis';
 export * from './target';
-export * from './skills';
-export * from './edu';
+export * from '../modules/skills/types';
+export * from '../modules/grad/types';
 export * from './app';
